Add MetaTags component and refactor layout components
