+++ components/Main.tsx
@@ -5,8 +5,9 @@ import { useReload } from "context/ReloadContext";
 import PoolList from "./PoolList";
 import PoolPage from "./PoolPage";
 import { Column, useIsMobile } from "utils/chakraUtils";
-import { Header } from "components/shared/Header";
-import Footer from "components/shared/Footer";
+import { Header } from "components/layouts/Header";
+import Footer from "components/layouts/Footer";
+import MetaTags from "components/layouts/MetaTags";
 
 const Main = () => {
   const { pageName } = useAppData();
@@ -27,19 +28,22 @@ const Main = () => {
   }, [address, reload]);
 
   return (
-    <Column
-      mainAxisAlignment="flex-start"
-      crossAxisAlignment="center"
-      color="#FFFFFF"
-      mx="auto"
-      width={isMobile ? "100%" : "1150px"}
-      px={isMobile ? 4 : 0}
-    >
-      <Header />
-      {pageName === "list" && <PoolList />}
-      {pageName === "pool" && <PoolPage />}
-      <Footer />
-    </Column>
+    <>
+      <MetaTags />
+      <Column
+        mainAxisAlignment="flex-start"
+        crossAxisAlignment="center"
+        color="#FFFFFF"
+        mx="auto"
+        width={isMobile ? "100%" : "1150px"}
+        px={isMobile ? 4 : 0}
+      >
+        <Header />
+        {pageName === "list" && <PoolList />}
+        {pageName === "pool" && <PoolPage />}
+        <Footer />
+      </Column>
+    </>
   );
 };
