+++ components/layouts/Footer.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 import { Link, Text } from "@chakra-ui/react";
 import { Row, Column } from "utils/chakraUtils";
-import CopyrightSpacer from "./CopyrightSpacer";
+import CopyrightSpacer from "components/shared/CopyrightSpacer";
 
 const Footer = () => {
   const { t } = useTranslation();
